+++ components/buttonText/styles.ts
@@ -1,6 +1,0 @@
-import { StyleSheet } from 'react-native';
-import { ITheme } from '../provider';
-
-export const styles = (config: ITheme) => StyleSheet.create({
-
-});