+++ components/index.ts
@@ -1,4 +1,7 @@
-export { default as ButtonText } from './buttonText';
 export { default as Button } from './button';
+export { default as TextField } from './textField';
+export { default as Checkbox } from './checkbox';
 export { default as Card } from './card';
+export { default as Link } from './link';
+export { default as Divider } from './divider';
 export { default as SavanaProvider } from './provider';