+++ components/provider/index.tsx
@@ -1,8 +1,13 @@
 import { ReactNode, createContext } from 'react';
+import { AnimatableNumericValue, DimensionValue, TextStyle } from 'react-native';
 
 type HexadecimalType = `#${string}`
 
+type Percentage = `${number}%`
+
 type StringOrNumber = string | number
+
+type FontWeight = Exclude<TextStyle['fontWeight'], 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900>
 
 export interface ITheme {
   colors: {
@@ -19,41 +24,40 @@ export interface ITheme {
     tertiaryLight?: HexadecimalType
     tertiary?: HexadecimalType
     tertiaryDark?: HexadecimalType
+    grayLight?: HexadecimalType
+    gray?: HexadecimalType
+    errorLight?: HexadecimalType
     error?: HexadecimalType
+    errorDark?: HexadecimalType
+    warningLight?: HexadecimalType
     warning?: HexadecimalType
+    warningDark?: HexadecimalType
+    successLight?: HexadecimalType
     success?: HexadecimalType
+    successDark?: HexadecimalType
+    infoLight?: HexadecimalType
     info?: HexadecimalType
+    infoDark?: HexadecimalType
   },
-  space?: Record<StringOrNumber, StringOrNumber>
-  borderWidths?: Record<StringOrNumber, StringOrNumber>
-  radii?: Record<StringOrNumber, StringOrNumber>
-  breakpoints?: Record<StringOrNumber, StringOrNumber>
-  letterSpacings?: Record<StringOrNumber, StringOrNumber>
-  lineHeights?: Record<StringOrNumber, StringOrNumber>
-  fontWeights?: {
-    hairline?: number
-    thin?: number
-    light?: number
-    normal?: number
-    medium?: number
-    semibold?: number
-    bold?: number
-    extrabold?: number
-    black?: number
-    extraBlack?: number
-  },
-  fonts?: {
+  space: Record<StringOrNumber, Percentage | number>
+  borderWidths: Record<StringOrNumber, number>
+  radii: Record<StringOrNumber, AnimatableNumericValue>
+  breakpoints: Record<StringOrNumber, DimensionValue>
+  letterSpacings: Record<StringOrNumber, DimensionValue>
+  lineHeights: Record<StringOrNumber, DimensionValue>
+  fontWeights: Record<StringOrNumber, Exclude<TextStyle['fontWeight'], 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900>>,
+  fonts: {
     heading?: string,
     body?: string,
     mono?: string,
   },
-  fontSizes?: Record<StringOrNumber, StringOrNumber>,
-  opacity?: Record<StringOrNumber, StringOrNumber>
+  fontSizes: Record<StringOrNumber, number>,
+  opacity: Record<StringOrNumber, DimensionValue>
 }
 
-const defaultTheme: ITheme = {
+export const defaultTheme = {
   colors: {
-    light: '#EEF4F2',
+    light: '#efeef1',
     dark: '#B7C0BC',
     white: '#FFFFFF',
     black: '#000000',
@@ -66,10 +70,20 @@ const defaultTheme: ITheme = {
     tertiaryLight: '#e0e4e3',
     tertiary: '#b7c0bc',
     tertiaryDark: '#8b928f',
+    grayLight: '#e1e1e1',
+    gray: '#686763',
+    errorLight: '#ff76764a',
     error: '#D84848',
+    errorDark: '#a83232',
+    warningLight: '#ffe0664a',
     warning: '#fdc416',
+    warningDark: '#b6912d',
+    successLight: '#75e17b4a',
     success: '#22c55e',
+    successDark: '#1d9a41',
+    infoLight: '#6ac7ff4a',
     info: '#0ea5e9',
+    infoDark: '#0b7ba3'
   },
   space: {
     0: 0,
@@ -97,31 +111,31 @@ const defaultTheme: ITheme = {
     128: 256,
     144: 288,
     160: 320,
-    '1/2': '50%',
-    '1/3': '33.333%',
-    '2/3': '66.666%',
-    '1/4': '25%',
-    '2/4': '50%',
-    '3/4': '75%',
-    '1/5': '20%',
-    '2/5': '40%',
-    '3/5': '60%',
-    '4/5': '80%',
-    '1/6': '16.666%',
-    '2/6': '33.333%',
-    '3/6': '50%',
-    '4/6': '66.666%',
-    '5/6': '83.333%',
-    '1/10': '10%',
-    '2/10': '20%',
-    '3/10': '30%',
-    '4/10': '40%',
-    '5/10': '50%',
-    '6/10': '60%',
-    '7/10': '70%',
-    '8/10': '80%',
-    '9/10': '90%',
-    'full': '100%',
+    '1/2': '50%' as Percentage,
+    '1/3': '33.333%' as Percentage,
+    '2/3': '66.666%' as Percentage,
+    '1/4': '25%' as Percentage,
+    '2/4': '50%' as Percentage,
+    '3/4': '75%' as Percentage,
+    '1/5': '20%' as Percentage,
+    '2/5': '40%' as Percentage,
+    '3/5': '60%' as Percentage,
+    '4/5': '80%' as Percentage,
+    '1/6': '16.666%' as Percentage,
+    '2/6': '33.333%' as Percentage,
+    '3/6': '50%' as Percentage,
+    '4/6': '66.666%' as Percentage,
+    '5/6': '83.333%' as Percentage,
+    '1/10': '10%' as Percentage,
+    '2/10': '20%' as Percentage,
+    '3/10': '30%' as Percentage,
+    '4/10': '40%' as Percentage,
+    '5/10': '50%' as Percentage,
+    '6/10': '60%' as Percentage,
+    '7/10': '70%' as Percentage,
+    '8/10': '80%' as Percentage,
+    '9/10': '90%' as Percentage,
+    'full': '100%' as Percentage,
   },
   borderWidths: {
     '0': 0,
@@ -173,16 +187,16 @@ const defaultTheme: ITheme = {
     '7xl': 90,
   },
   fontWeights: {
-    hairline: 100,
-    thin: 200,
-    light: 300,
-    normal: 400,
-    medium: 500,
-    semibold: 600,
-    bold: 700,
-    extrabold: 800,
-    black: 900,
-    extraBlack: 950,
+    hairline: '100' as FontWeight,
+    thin: '200' as FontWeight,
+    light: '300' as FontWeight,
+    normal: '400' as FontWeight,
+    medium: '500' as FontWeight,
+    semibold: '600' as FontWeight,
+    bold: '700' as FontWeight,
+    extrabold: '800' as FontWeight,
+    black: '900' as FontWeight,
+    extraBlack: '950' as FontWeight,
   },
   fonts: {
     heading: undefined,
@@ -224,6 +238,8 @@ const defaultTheme: ITheme = {
   }
 }
 
+export type DefaultTheme = typeof defaultTheme
+
 const mergeObject = (target: ITheme, source: Partial<ITheme>): ITheme => {
   const isObject = (obj: any) => obj && typeof obj === 'object';
 
@@ -243,10 +259,10 @@ const mergeObject = (target: ITheme, source: Partial<ITheme>): ITheme => {
   }, { ...target });
 };
 
-export const ContextTheme = createContext(defaultTheme);
+export const ContextTheme = createContext(defaultTheme as ITheme);
 
 export const SavanaProvider: React.FC<{ config?: ITheme; children: ReactNode }> = ({ config, children }): JSX.Element => {
-  const currentTheme = mergeObject(defaultTheme, config ?? {})
+  const currentTheme = mergeObject(defaultTheme as ITheme, config ?? {})
 
   return (
     <ContextTheme.Provider value={currentTheme}>